move uuid base to own module
