+++ src/cript/nodes/primary_nodes/reference.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field, replace
 from typing import List
 
-from cript.nodes.primary_nodes.primary_base_node import UUIDBaseNode
+from cript.nodes.uuid_base import UUIDBaseNode
 
 
 class Reference(UUIDBaseNode):
